+++ src/RecommendPage/RecommendListBody.js
@@ -1,26 +1,22 @@
 import React from 'react'
-import "../pages/FeaturedListBody.css"
-import Header from '../Header';
-import SongRow from "../SongRow";
+import "../RecommendPage/RecommendListBody.css"
+import Header from '../Header/Header';
+import SongRow from "../SongPlay/SongRow";
 import { useDataLayerValue } from "../DataLayer";
 import PlayCircleFilledIcon from '@material-ui/icons/PlayCircleFilled';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
 
-function SearchPlaylist({spotify}) {
-  const [{ searchplaylist } , dispatch] = useDataLayerValue();
-  console.log("searchplaylist",)
+
+function FeaturedListBody({spotify}) {
+  const [{ featuredtracks } , dispatch] = useDataLayerValue();
 
   const playPlaylist = () => {
-    
-
     spotify.play({
-      context_uri: `spotify:playlist:${searchplaylist?.id}`,
+      context_uri: `spotify:playlist:${featuredtracks?.id}`,
     })
-    
     .then((res) => {
       spotify.getMyCurrentPlayingTrack().then((response) => {
-
         dispatch({
           type:"SET_ITEM",
           item: response.item,
@@ -53,18 +49,19 @@ function SearchPlaylist({spotify}) {
   };
   
 
+
   return (
     <div className="FeaturedListBody"  >
       <Header spotify={spotify} />
       <div className="body__info">
       <img 
-          src={searchplaylist?.images[0].url}
+          src={featuredtracks?.images[0].url}
           alt=""
           />
        <div className="body__infoText">
           <strong>PLAYLIST</strong>
-          <h2>{searchplaylist?.name}</h2>
-          <p>{searchplaylist?.description}</p>
+          <h2>{featuredtracks?.name}</h2>
+          <p>{featuredtracks?.description}</p>
         </div>
       </div>
       <div classname="body__songs">
@@ -76,15 +73,15 @@ function SearchPlaylist({spotify}) {
           <FavoriteIcon fontSize="large"/>
           <MoreHorizIcon />
         </div>
-          {searchplaylist?.tracks.items.map((item) => (
+          {featuredtracks?.tracks.items.map((item) => (
             <SongRow 
             playSong={playSong} 
              track={item.track} 
              />
             ))}
       </div>
-    </div> 
+    </div>
   )
 }
 
-export default SearchPlaylist
+export default FeaturedListBody
