+++ src/SearchPage/SearchResult.css
