+++ src/spotify.js
@@ -5,8 +5,8 @@ export const authEndpoint =
 
 
 // // Spotify for Developerから引用
-const redirectUri = "https://spotify-app-pasha1235.web.app/";
-// const redirectUri = "http://localhost:3000/";
+// const redirectUri = "https://spotify-app-pasha1235.web.app/";
+const redirectUri = "http://localhost:3000/";
 
 
 const clientID = "f7b768418c9a4246816351ba46e0b52f";
